+++ tests/testthat/test-file-cache.R
@@ -4,9 +4,6 @@ test_that("raadtools config works", {
   # oldpath <- getOption("default.datadir")
   # expect_true(file.exists(oldpath))
 
-  newpath <- getOption("raadfiles.default.data.directory")
-  expect_true(oldpath == newpath)
-
   db <- getOption("raadfiles.filename.database")
   db %>% expect_s3_class("tbl_df") %>% expect_named(c("root", "file"))